refactor: move backupstore fucntions to separate file
